Split express config and routes into app.js with server.ts only starting server
